# SwaggerDoc/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Member Services",
    "version" : "0.1",
    "title" : "Member Services"
  },
  "paths" : {
    "/members/" : {
      "get" : {
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Member Array in JSON format",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Member"
              }
            }
          },
          "404" : {
            "description" : "No member exists in the system.",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ResponseMessage"
              }
            }
          },
          "default" : {
            "description" : "Get all members"
          }
        }
      },
      "post" : {
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Member"
          },
          "x-examples" : {
            "application/json" : "{\"firstName\":\"James\",\"lastName\":\"Bond\",\"email\":\"jbond@mi6.gov.uk\"}"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "ResponseMessage",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ResponseMessage"
              }
            }
          },
          "412" : {
            "description" : "Invalid first name, last name and/or email [OR] duplciate member name or email",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ResponseMessage"
              }
            }
          },
          "default" : {
            "description" : "Add a new member"
          }
        }
      }
    },
    "/members/replaceWithMemberIdToSearch" : {
      "get" : {
        "parameters" : [ {
          "name" : "memberId",
          "in" : "path",
          "description" : "Numeric ID of the member to get",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "Member",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Member"
              }
            }
          },
          "404" : {
            "description" : "Member not found.",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ResponseMessage"
              }
            }
          },
          "default" : {
            "description" : "Get a member by ID"
          }
        }
      }
    },
    "/members/replaceWithMemberIdToDelete" : {
      "delete" : {
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "name" : "memberId",
          "in" : "path",
          "description" : "Numeric ID of the member to delete",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "ResponseMessage",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ResponseMessage"
              }
            }
          },
          "404" : {
            "description" : "Member not found.",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ResponseMessage"
              }
            }
          },
          "default" : {
            "description" : "Delete a member by ID"
          }
        }
      }
    },
    "/members/replaceWithMemberIdToUpdate" : {
      "patch" : {
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "name" : "memberId",
          "in" : "path",
          "description" : "Numeric ID of the member to update",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Member"
          },
          "x-examples" : {
            "application/json" : "{\"firstName\":\"James\",\"lastName\":\"Bond\",\"email\":\"jbond@mi6.gov.uk\"}"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "ResponseMessage",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ResponseMessage"
              }
            }
          },
          "404" : {
            "description" : "Member not found.",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ResponseMessage"
              }
            }
          },
          "412" : {
            "description" : "Invalid first name, last name and/or email [OR] duplciate member name or email",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ResponseMessage"
              }
            }
          },
          "default" : {
            "description" : "Update a member first name, last name, email by ID"
          }
        }
      }
    }
  },
  "definitions" : {
    "Member" : {
      "properties" : {
        "firstName" : {
          "type" : "string"
        },
        "lastName" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        }
      },
      "example" : {
        "firstName" : "firstName",
        "lastName" : "lastName",
        "email" : "email"
      }
    },
    "ResponseMessage" : {
      "properties" : {
        "message" : {
          "type" : "string"
        }
      },
      "example" : {
        "message" : "message"
      }
    }
  }
}